+++ Models/SSB/AgesInNorway.cs
@@ -1,10 +1,12 @@
 using System.Text.Json.Serialization;
+using Backend.Models.JsonStat;
 
-namespace Backend.Models.JsonStat
+namespace Backend.Models.SSB
 {
     // Remove comments when merge-ready
-    // Based on https://data.ssb.no/api/v0/dataset/1074.json?lang=no
-    public class SSBAgesInNorway
+    // Based on https://data.ssb.no/api/v0/dataset/26975.json?lang=no
+    // https://data.ssb.no/api/v0/dataset/26975?lang=no
+    public class AgesInNorway
     {
         /*
          * I tried abstracting this class as much as possible, so that the only thing
@@ -13,18 +15,14 @@ namespace Backend.Models.JsonStat
          */
         [JsonPropertyName("dataset")] public JsonStatDataset Dataset { get; set; }
 
-        public sealed class JsonStatDataset : JsonStatGenericDataset
+        public sealed class JsonStatDataset : AbstractJsonStatDataSet
         {
             [JsonPropertyName("dimension")] public JsonStatSpecificDimension Dimension { get; set; }
         }
 
-        public sealed class JsonStatSpecificDimension : JsonStatGenericDimension
+        public sealed class JsonStatSpecificDimension : AbstractJsonStatDimension
         {
-            /*
-             * Changes are done here, the rest _should_ be equivalent for other json-stat objects
-             */
             [JsonPropertyName("Region")] public JsonStatDimensionContent Region { get; set; }
-            [JsonPropertyName("Alder")] public JsonStatDimensionContent Alder { get; set; }
             [JsonPropertyName("Tid")] public JsonStatDimensionContent Tid { get; set; }
         }
     }
